Updated wequire to support a mode that overwrites all existing files, and tentatively added support for multiple URLs

+++ lib/wequire.lua
@@ -1,23 +1,28 @@
 -- local baseUrl = should be injected by /bootstrap.lua
+local config = require("/lib/config").load("wequire")
 
 local m = {}
 
 -- cache functions in case these global functions are overwritten with these versions
-local require = require
-local loadfile = loadfile
+local _require = require
+local _loadfile = loadfile
+
+m.overwrite = false
 
 m.fetch = function(file)
-  local response = http.get({ url = baseUrl .. file, binary = true })
+  for _, url in ipairs(config) do
+    local response = http.get({ url = url .. file, binary = true })
 
-  if response and response.getResponseCode() == 200 then
-    if fs.exists(file) then
-      error("Failed to download '" .. file .. "': file already exists.")
-    else
-      fs.open(file, "wb").write(response.readAll())
+    if response and response.getResponseCode() == 200 then
+      if not m.overwrite and fs.exists(file) then
+        error("Failed to write '" .. file .. "': file already exists.")
+      else
+        fs.open(file, "wb").write(response.readAll())
 
-      print("Fetched " .. file .. " from " .. baseUrl .. file)
+        print("Fetched " .. file .. " from " .. url .. file)
 
-      return true
+        return true
+      end
     end
   end
 
@@ -25,26 +30,26 @@ m.fetch = function(file)
 end
 
 m.require = function(file)
-  
-  if _ENV.isWerun then fs.delete(file..".lua") end
+  if not m.overwrite then
+    local ok, val = pcall(_require, file)
 
-  local ok, val = pcall(require, file)
-
-  if ok then return val end
+    if ok then return val end
+    print(val)
+  end
 
   for search in package.path:gmatch("[^;]+") do
     local f = search:gsub("?", file)
 
     if m.fetch(f) then
-      return require(file)
+      return _require(file)
     end
   end
 
   error("Unable to find file locally or online")
 end
 
-m.loadfile = function(file)
-  local f = loadfile(file)
+m.loadfile = function(file, mode, env)
+  local f = _loadfile(file, mode, env)
 
   if f then return f end
 
@@ -53,36 +58,38 @@ m.loadfile = function(file)
   fetchList[#fetchList + 1] = file
 
   for p in shell.path():gmatch("[^:]+") do
-    for _, pp in ipairs({ file, file .. ".lua"}) do
-      local ppp = p .. "/" .. pp
+    if fs.getDrive(p) ~= "rom" then
+      for _, pp in ipairs({ file, file .. ".lua"}) do
+        local ppp = p .. "/" .. pp
 
-      if _ENV.isWerun and string.sub(ppp, 1, 5) ~= "/rom/" then
-        fs.delete(ppp)
+        if not m.overwrite then
+          f = _loadfile(ppp, mode, env)
+
+          if f then return f end
+        end
+
+        fetchList[#fetchList + 1] = ppp
       end
-
-      f = loadfile(ppp)
-
-      if f then return f end
-
-      fetchList[#fetchList + 1] = ppp
     end
   end
 
   for _, fet in ipairs(fetchList) do
     if m.fetch(fet) then
-      return loadfile(fet)
+      return _loadfile(fet, mode, env)
     end
   end
 
-  error("Unable to find file locally or online")
+  return nil, "Unable to find file '" .. file .. "' locally or online"
 end
 
-m.run = function(environment, file, ...)
-  local r = m.loadfile(file)
+m.run = function(env, file, ...)
+  local f, err = m.loadfile(file, "bt", env)
 
-  setfenv(r, environment)
-
-  r(...)
+  if f then
+    return f(...)
+  else
+    error(err)
+  end
 end
 
 return m
